+++ NEWS.md
@@ -1,8 +1,9 @@
 # redistmetrics 1.0.1
 
-* Fixed bug where `part_bias()` was reversed in sign.
-* Fixed bug where `splits_admin()` plans input needed to be sequentially numbered (#10).
+* Fixes bug where `part_bias()` was reversed in sign.
+* Fixes bug where `splits_admin()` plans input needed to be sequentially numbered (#10).
 * Allows planarizing warnings to be silenced when `epsg = FALSE` (#11).
+* Makes `ncores` argument work for compactness functions.
 
 # redistmetrics 0.0.0.9000